consolidate auth into a single component
