+++ spotlight-client/src/AuthWall/utils.test.ts
@@ -17,7 +17,7 @@ afterEach(() => {
  * for your test, if any.
  */
 async function getGetAuthSettings() {
-  return (await import("./getAuthSettings")).default;
+  return (await import("./utils")).getAuthSettings;
 }
 
 test("returns nothing when the value is unset", async () => {
@@ -46,6 +46,33 @@ test("returns a settings object when the value is supported", async () => {
   // just that something of the valid type is provided
   expect(typeof settings.domain).toBe("string");
   expect(typeof settings.clientId).toBe("string");
+});
+/**
+ * Dynamically imports the isAuthEnabled for testing.
+ * Should be used after setting the desired environment variables
+ * for your test, if any.
+ */
+async function getIsAuthEnabled() {
+  return (await import("./utils")).isAuthEnabled;
+}
+
+test("returns false when the value is unset", async () => {
+  const isAuthEnabled = await getIsAuthEnabled();
+  expect(isAuthEnabled()).toBe(false);
+});
+
+test("returns false when the value is not 'true'", async () => {
+  process.env.REACT_APP_AUTH_ENABLED = "false";
+
+  const isAuthEnabled = await getIsAuthEnabled();
+  expect(isAuthEnabled()).toBe(false);
+});
+
+test("returns true when the value is 'true'", async () => {
+  process.env.REACT_APP_AUTH_ENABLED = "true";
+
+  const isAuthEnabled = await getIsAuthEnabled();
+  expect(isAuthEnabled()).toBe(true);
 });
 
 // this doesn't do anything except appease Typescript;
